Fix logic validations not being applied to ego variables
Two validation rules were not being applied to ego variables.

  - Validation of variable validation that references other variables
  (sameAs, differentFrom, greaterThanVariable, lessThanVariable)
  - Validation that variable name is unique

Both issues are fixed here by using a regexp directly to specify the
validation target, and then adding specific logic to deal with parsing
the ego codebook.

This change technically represents a breaking change within the schema,
but in practice we have no bandwidth to properly handle the migration,
and very little reason to believe that protocols exist that are broken
due to this issue.